[Bug]Only call getRangeAt(0) if rangeCount > 0 (#545)
From what I understand the browser asynchronously updates the selection object and so there are scenarios where the rangeCount is not updated by the time we query it. We previously were blindly calling getRangeAt(0) but there were no ranges. This is like trying to access element on array which is empty. We now check the length first.

https://developer.mozilla.org/en-US/docs/Web/API/Selection/getRangeAt

Fixes VSTS#1333