test: Add Train Dialogs test. (#598)
* test: adding train dialog test

Once the new model has been created, this test step validates the creation of a Train Dialog, and
then validates if the messages from the user and the bot are displayed in the chat pannel. Also,
this code change adds data-testids in preparation for new tests. Finally, the change also includes
the use of Components.

* test: fixing support files configuration

fixing the way how the existance of support files is being resolved by cypress

* test: correcting entity test step

adding a second between entity related steps

* test: changing entity test which is still failing

changing entity test which is still failing by removing second entity creation

* test: adding AUT components

initial creation of application under test components with the functions needed only to meet the
hello world test, more functions will be added as they are required.