fix: prevent unnecessary poll if deleting new teach session (#930)
We were polling for train status even when it's known that the status should not have changed. This might mislead people into thinking they caused changes and not know what they are.

It's a rare case because it only happens if you open teach session and close it before you select an action which triggers polling anyway