Fix Validations that Occur after Saving a Train Dialog (#1067)
After saving a Train Dialog there are times that the rendering takes a long time or it oscilates such that capturing it in the final settled state is difficult to achieve. I've filed bugs on these, I hope they will one day be fixed, but until then these changes should alow us to capture that final state with a very high degree of confidence and verify that it is as it is expected to be.

However, because of this change tests won't fail under these circumstances and we won't know that rendering is odd or taking too long...unless we search the logs for the issues.

Refer to bugs 1950, 1916 & 1855 for more info on the issues.