Fix Delete All Test Models test case (#1141)
We use a Build Key that recycles every 26 builds, which gives us an effective 13 unique keys before they are reused. This is plenty to ensure we don't reuse the same key.

We delete based on the key, if the model name has the Build Key in it that matches the current build we always delete it. If not we delete the model if the created time is 5 minutes back or longer.