Test Grid Updates and Fix for SaveAsIsVerifyInGrid (#1149)
Updates to the TestGrid.xlsx - For all covered test scenarios the test spec files are listed

SaveAsIsVerifyInGrid Fix - This fix is to deal with the rendering issues that come up when a Train Dialog is saved. While I am not happy with how convoluted this function and VerifyTrainingSummaryIsInGrid are becoming, right now I don't have a clear idea of the perfect fix. So I want to check it in and run with these changes for now, and see if any other rendering issues cause test failures while I think about how to write a better fix for this.