Delete All Test Generated Models (#1317)
Problem: Test Entity Names are being created in a NON-standard way. Example: "z-entityPicker2-09-6-45-32" is what I am seeing, it is missing the single character suffix use by the delete algorithm. The result is that these models were not being deleted and eventually they caused other test failures.

Solution: I also modified the "DeleteAllTestGeneratedModels" so that it will always delete a "z-" prefixed test model if the rest of the name is in a non-standard format. Previously I was not deleting those since I could not determine which test run it belonged to. The consequence of this change is that it is now possible (but still unlikely) that a test will fail because the model it is using will be deleted by another test run only if it is using a non-standard naming convention.