Fix Bug in cy.ExactMatch Command
I had changed the cy.ExactMatch command to behave more like the cy.contains command in terms of the number of elements it returns, however this caused too many tests to fail and the fix was necessary for only one function. So I worked out how to fix the issue for that one function without affecting all of the other tests.