Fix comprehensive1 smoke test (#22)
The Comprehensive1 Smoke test suite had a hack-around a UI bug. The bug has been fixed so I fixed test by commenting out the hack-around...commented rather than deleted in case the bug regresses then the code to deal with it is just a few comments.

The data that controls the test also had to be changed.